Recoded the Tmdb_Helper Function.
Ok so on my previous change, i added tvdb_id and tvdb_season to the database to help me do conversions and on tmdb, i found out you can get the tvdb series id from tvdb which is great so i coded this whole long function for a day and i manage to get Movies, TV Shows and Seasons and Parts too.

One problem i encountered was that embeds were not working on Movies and Parts and i was not sure why,

@Gujal00 , I coded this but i used the Requests Module and used python 3 in this code, if you could look into this and probably make completable with python 2 and change request to client request sense i have no idea how to work with client request, that would be awesome.

I hope to improve on this more in the future.
Thank you