Finally Fixed Local File Playback
It's been a few days and i have been working on recoding the whole database sense it was very messy. Apparently the URL Error was apparently due to due to _prefetch_play_link not having a file link and this was hard to find cause it was not specified in the error log.

Finally happy i got it, all i have is py2 conversions and i am finally done with this and will be ready to release.

Last Thing to do:
- Py2 Conversions.